Reorganize server and service test runner
Ref #2698
