+++ lib/request-handler.js
@@ -3,19 +3,19 @@
 // eslint-disable-next-line node/no-deprecated-api
 const domain = require('domain')
 const request = require('request')
-const { makeBadgeData: getBadgeData } = require('./badge-data')
-const log = require('./log')
+const queryString = require('query-string')
+const log = require('../core/server/log')
+const analytics = require('../core/server/analytics')
 const LruCache = require('../gh-badges/lib/lru-cache')
 const makeBadge = require('../gh-badges/lib/make-badge')
-const analytics = require('./analytics')
-const { makeSend } = require('./result-sender')
-const queryString = require('query-string')
 const {
   Inaccessible,
   InvalidResponse,
   ShieldsRuntimeError,
 } = require('../services/errors')
 const { setCacheHeaders } = require('../services/cache-headers')
+const { makeBadgeData: getBadgeData } = require('./badge-data')
+const { makeSend } = require('./result-sender')
 
 // We avoid calling the vendor's server for computation of the information in a
 // number of badges.
