+++ lib/request-handler.spec.js
@@ -4,8 +4,8 @@ const { expect } = require('chai')
 const fetch = require('node-fetch')
 const nock = require('nock')
 const portfinder = require('portfinder')
-const analytics = require('./analytics')
 const Camp = require('camp')
+const analytics = require('../core/server/analytics')
 const { makeBadgeData: getBadgeData } = require('./badge-data')
 const {
   handleRequest,
