+++ lib/result-sender.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const stream = require('stream')
-const log = require('./log')
+const log = require('../core/server/log')
 const svg2img = require('../gh-badges/lib/svg-to-img')
 
 function streamFromString(str) {
