+++ package.json
@@ -80,14 +80,14 @@
     "test:js:frontend": "echo \"Deprecated; run `npm run test:frontend` instead.\" && npm run test:frontend",
     "test:js:package": "echo \"Deprecated; run `npm run test:package` instead.\" && npm run test:package",
     "test:frontend": "cross-env BABEL_ENV=test mocha --opts mocha.opts --require @babel/polyfill --require @babel/register --require mocha-yaml-loader \"frontend/**/*.spec.js\"",
-    "test:server": "cross-env NODE_CONFIG_ENV=test mocha --opts mocha.opts \"lib/**/*.spec.js\" \"services/**/*.spec.js\"",
+    "test:server": "cross-env NODE_CONFIG_ENV=test mocha --opts mocha.opts \"core/**/*.spec.js\" \"lib/**/*.spec.js\" \"services/**/*.spec.js\"",
     "test:package": "mocha --opts mocha.opts \"gh-badges/**/*.spec.js\"",
     "test:entrypoint": "cross-env NODE_CONFIG_ENV=test mocha --opts mocha.opts entrypoint.spec.js",
     "test:integration": "cross-env NODE_CONFIG_ENV=test mocha --opts mocha.opts \"lib/**/*.integration.js\" \"services/**/*.integration.js\"",
-    "test:services": "cross-env NODE_CONFIG_ENV=test mocha --opts mocha.opts --delay lib/service-test-runner/cli.js",
+    "test:services": "cross-env NODE_CONFIG_ENV=test mocha --opts mocha.opts --delay core/service-test-runner/cli.js",
     "test:services:trace": "cross-env NODE_CONFIG_ENV=test TRACE_SERVICES=true npm run test:services -- $*",
-    "test:services:pr:prepare": "node lib/service-test-runner/pull-request-services-cli.js > pull-request-services.log",
-    "test:services:pr:run": "cross-env NODE_CONFIG_ENV=test mocha --opts mocha.opts --delay lib/service-test-runner/cli.js --stdin < pull-request-services.log",
+    "test:services:pr:prepare": "node core/service-test-runner/pull-request-services-cli.js > pull-request-services.log",
+    "test:services:pr:run": "cross-env NODE_CONFIG_ENV=test mocha --opts mocha.opts --delay core/service-test-runner/cli.js --stdin < pull-request-services.log",
     "test:services:pr": "run-s --silent test:services:pr:prepare test:services:pr:run",
     "pretest": "run-s --silent defs features",
     "test": "run-s --silent --continue-on-error lint test:frontend test:package test:server test:entrypoint prettier-check",
