+++ services/travis/travis-php-version.service.js
@@ -2,7 +2,7 @@
 
 const LegacyService = require('../legacy-service')
 const { makeBadgeData: getBadgeData } = require('../../lib/badge-data')
-const log = require('../../lib/log')
+const log = require('../../core/server/log')
 const {
   minorVersion: phpMinorVersion,
   versionReduction: phpVersionReduction,
